refactor(e2e): move shared team helper functions to central helper file
